requirements.txt
pypdf
pinecone
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-ollama
langchain-openai
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langchain_pinecone
fastapi
uvicorn
python-dotenv
sentence-transformers
